refactor(web): lift conversation-store to src/stores/ (LUM-2017)
`useConversationStore` lived in `domains/conversations/` but is read or
written by every chat-adjacent domain (chat, messaging, conversations,
onboarding, page routes, sync). Per `docs/CONVENTIONS.md` — "Top-level
shared directories" — state used by two or more domains belongs at the
top level.

Move:
- src/domains/conversations/conversation-store.ts → src/stores/conversation-store.ts
- src/domains/conversations/conversation-store.test.ts → src/stores/conversation-store.test.ts

Codemod all imports (27 production files) from
`@/domains/conversations/conversation-store` →
`@/stores/conversation-store`.

Regenerate `.cross-domain-allowlist.json` via `bun run audit:cross-domain`:
- 13 chat→conversations allowlist entries deleted
- 3 files (use-refresh-latest-messages.ts + .test.ts, queue-handlers.ts)
  drop off the allowlist entirely — they only crossed the boundary for
  conversation-store

Update the store's header docstring to explain why it now lives at the
top level. No behavioural change — pure structural refactor.

Resolves the bulk of LUM-1922 (chat ↔ conversations bidirectional
coupling) and unblocks LUM-2019 (turn-termination coordinator, which
needs a top-level home).
