+++ apps/web/src/domains/chat/hooks/use-conversation-history.ts
@@ -39,7 +39,7 @@ import {
 } from "@/domains/chat/utils/diagnostics";
 import type { TranscriptPaginationState } from "@/domains/chat/transcript/types";
 import type { ContextWindowUsage } from "@/domains/chat/components/context-window-indicator";
-import { useConversationStore } from "@/domains/conversations/conversation-store";
+import { useConversationStore } from "@/stores/conversation-store";
 import { useInteractionStore } from "@/domains/interactions/interaction-store";
 import { useSubagentStore } from "@/domains/subagents/subagent-store";
 import type { SubagentStatus } from "@/domains/chat/api/event-types";
