+++ apps/web/src/domains/chat/utils/stream-handlers/metadata-handlers.ts
@@ -8,7 +8,7 @@ import {
 import { patchConversation } from "@/domains/conversations/conversation-queries";
 import type { StreamHandlerContext } from "@/domains/chat/utils/stream-handlers/types";
 import type { AvatarUpdatedEvent, CompactionCircuitClosedEvent, CompactionCircuitOpenEvent, ConversationTitleUpdatedEvent, DiskPressureStatusChangedEvent, IdentityChangedEvent, NotificationIntentEvent, TurnProfileAutoRoutedEvent, UsageUpdateEvent } from "@/domains/chat/api/event-types";
-import { useConversationStore } from "@/domains/conversations/conversation-store";
+import { useConversationStore } from "@/stores/conversation-store";
 
 export function handleUsageUpdate(
   event: UsageUpdateEvent,
