+++ apps/web/src/domains/conversations/use-attention-tracking-mark-seen.test.tsx
@@ -18,7 +18,7 @@ import type { ReactNode } from "react";
 import { createElement } from "react";
 
 import * as sdkGen from "@/generated/daemon/sdk.gen";
-import { useConversationStore } from "@/domains/conversations/conversation-store";
+import { useConversationStore } from "@/stores/conversation-store";
 import { __resetEventBusForTesting } from "@/stores/event-bus-store";
 
 // ---------------------------------------------------------------------------
