+++ apps/web/src/identity-page-route.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from "react-router";
 
 import { useActiveAssistantContext } from "@/components/layout/active-assistant-gate";
 import { createDraftConversationId } from "@/domains/chat/utils/conversation-selection";
-import { useConversationStore } from "@/domains/conversations/conversation-store";
+import { useConversationStore } from "@/stores/conversation-store";
 import { IdentityPage } from "@/domains/intelligence/identity-page";
 import { useViewerStore } from "@/stores/viewer-store";
 import { routes } from "@/utils/routes";
